// src/modules/products/interfaces/requests/update-product-request.interface.ts
import { ProductCategory } from "@/modules/products/enums/product-category.enum";
import { ProductCondition } from "@/modules/products/enums/product-condition.enum";
import { PreferenceType } from "@/modules/client/enums/preference-type.enum";

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  stock?: number;
  originalPrice?: number;
  discountPercentage?: number;
  category?: ProductCategory;
  conditions?: ProductCondition[];
  expirationDate?: string;
  preferences?: PreferenceType[];
}